Reconcile mismatched Contact and To/From URIs early.
Fixes dropped state for caller URI (and stuck state for callee if callee
is SCA) on second pickup on different handset after hold. Depending on
the values in the To/From headers won't work, since the SCA reINVITE to
seize a held call uses the SCA group's AoR in RURI, To and From, leading
to problems in SCA-to-SCA calls.

This does not yet resolve the display URI problem.