modules/tm: Set branch_index to T_BR_UNDEFINED when outside BRANCH_ROUTE or TM_ONREPLY_ROUTE.
The inconsistent value of $T_branch_idx between BRANCH_ROUTE and
TM_ON_REPLY_ROUTE was fixed in an earlier commit, but now the value 0 has a
double meaning (branch 0 or invalid branch). This patch makes the invalid
branch distinguishable by setting it to -1.

Now $T_branch_idx will return the branch number (0-based) in BRANCH_ROUTE
and TM_ON_REPLY_ROUTE and -1 in other route types or if the message is not
part of a transaction.