modules_k/db_sqlite: fix crash with computed fields in custom queries
Computed fields do not have decltype available, so guess the proper
field type based on the result type of the first row. This does not
work if the first row has null type as result, but is the best we can
do easily and fixes gives right result in most cases.

Reported-by: Pedro Antonio Vico Solano <pvsolano@amper.es>