core: make database version table name configurable from script
When kamailio is part of a bigger project and is sharing its database with
it, it is nice to have a less generic table name than "version".

Also usefull when upgrading kamailio, where the new version has (backwards
compatibel) changes in the database definition. If each kamailio version uses
its own version table, they can all use the same database.