snmpstats Add tcpasync and tcpmaxconns
This is to test if I can reach core configuration settings for TCP. Tests prove that it's
possible. Will go ahead and add other settings as well.

One question is if we should allow changing these variables in SNMP, like we do
over the RPC interface or selects. Let's think about that.