snmpstats Add support for the Websocket module statistics
Hint to developers: If you add statistics and selects in your module, you
make it very easy to add SNMP support for your module. Then SNMPstats
can just use generic APIs to find your data. Please also add a generic
way of finding out if a module is configured and used too, if possible.
Sometimes modules are just loaded, but not used anywhere. Better to
reflect real data in SNMP if possible.

Using the old "SER" way of using selects for config data that
is allowed to change at runtime opens up for SNMP writes to manage
your module as well.

And of course, using selects and counters/statistics variables also
helps the RPC interface, so it's both cool and the Right Thing To Do (TM).

Yes, commit messages can be informative as well. :-)