call_control: Passthrough sip_application
Current patches being submitted to CDRTool and the call_control daemon
allow for SIP application type to be specified. This allows audio or
video services to be rated differently, or even different ratings for
application subtypes (audio.inbound vs. audio.outbound).

This patch allows the sip_application value to be passed through to the
call_control daemon.