dialog: Set the dialog context on incoming replies
Before, the dialog context would be derived from the callid and tags. This
could select the wrong dialog in case of spirals. This patch sets the correct
dialog context by storing the it in the transaction and restoring it on an
incoming response. Now the correct dialog variables are available in
onreply_route.

(cherry picked from commit 25a3297915fbcc0a267d444d245526de978a5d62)