dialog: Add syncing via DMQ
Primary purpose is to sync profiles. This implements just that.
Enough of the state is synced to maintain identical profiles counts on
all kamailio instances.

Changed profiles are only synced as the request is sent out, adding
multiple profiles will result in just 1 sync message at the moment
the request is being sent.

It is not possible for the 'other' instances to send in-dialog requests.