core: sip_msg_shm_clone: clone parsed maxforwards value in shm
maxfwd module modified the msg buf directly instead of adding a lump. In
addition it is filling the parsed value of the header field.

Clone this field into shm so later code can at least determine if the msg
has been mangled so it can retreive the original value of the header.