tm: Fix setting flags for record-route headers
The flags where set in add_uac() by counting the number of record-route
headers added by prepare_new_uac(). The latter function backups and
restores the list of lumps, so add_uac() only sees the original list,
which is not what has happened on the branch being handled.

Moving setting the flags into prepare_new_uac() fixes this.

This fixed the dialog module, which has to remove the correct number
of RR headers from the callee's routeset (which it does based on the
tm flags).