rtpengine: changes for mi_fifo_commands
Added a new fifo command, namely nh_ping_rtpp, which pings the given proxy or
all proxies, and if a proxy does not respond, disable it temporarily.
If the proxy responds, it does nothing.
Changed the existing fifos to execute for a specific proxy or for all proxies
depending on the first given parameter.
Changed nh_enable_rtpp to try to ping before enabling and enable
it only if ping success. Changed some formatting for nh_show_rtpp to
print 'N/A' when disabled permanent and to display the correct recheck_ticks.
Updated doku accordingly.