srdb1  Check if the module is loaded, so the error message makes sense in case of missing module
If a module use a default database and it's not loaded, currently
Kamailio complains that the module doesn't implement an API function
instead of actually telling the Kamailian that the module is not loaded.
This patch modifies that behaviour.

(cherry picked from commit c1a1a1bb054ee79af6ca44d262c8f73c725a2d90)