mem/q_malloc: log more diagnostic information in qm_debug_frag
I had an instance of SER crash in production with this error message from
qm_debug_frag:

[mem/q_malloc.c:149]: BUG: qm_*: fragm. 0x4cee3b88 (address 0x4cee3ba0) end
overwritten(0, 0)!

The fragment end was been overwritten with zeroes... buffer overflow? Or an
aliasing bug? Maybe a double-free?

It would be much easier to debug problems like this if qm_debug_frag printed the
call site where qm_malloc/free was invoked from. And now it does!