permissions: possibility to specify time period in routing rules
Possibility to specify a time period following the RFC 2445 (explained in the TMREC module readme)
at the end of a rule in the allow and deny files. The rule will be valid only during this period.
Usefull for company or people that don't want calls during night or want to receive calls from someone
only during a specific period of time. It was already possible with the TMREC module in a routing
script but this was redundant with the permission module so it simplifies the conf.