rtpengine: Allow op for all deactivated machines
If allow_op modparam enabled, send commands to the disabled machines for the
existing call. So far this was done only for manually deactivated machines.
This is useful because there might be cases of proxy timeout, cases when you
may want to still allow the operations for the existing calls.