rtpengine: Add hiding of deleted table nodes
'kamctl fifo nh_show_rtpp all' reflects the rtpengine table state.

When node is deleted from the table the node itself isn't freed but disabled
permanent and hidden for display. This is mainly because one might want the
current session to finish for the deleted table nodes (see allow_op modparam).

Hiding the nodes and not freeing them will spare time deciding if there are
any sessions left for the deleted rtpengine.