rr: append_fromtag on upstream in-dialog requests
Record-Routing for in-dialog request are not needed by RFC 3162. But
there are many clients out there not sending initial route-set in
in-dialog requests. To provided them with the initial route-set together
with UAC modified requests the ;ftag parameter must be the to-tag on
upstream requests.

append_fromtag parameter added the tag value from "from" header  always,
even on upstream. This is fixed here.

It is not done on record_route_preset(). This should not be called on
record routing anyway.