Free database column name as part of database result cleanup
The column name is currently not free'd. Some db bankends copy this
data so that memory is leaked. Some store internal database pointers
and those shouldn't be free'd. One returns a pointer to a stack
variable which shouldn't be done.

The patch cleans up all db backends to copy the column name and frees
that column name as part of the database result cleanup function.

Author: Chris Double <doublec@silentcircle.com>