auth: Add option to use SHA-256 instead of MD5 in digest authentication
The module paramter "algorithm" was added with the possible values "", "MD5" or "SHA-256".
It allows chosing between MD5 or SHA-256 hash algorithm for digest authentication.
The configured algorithm is advertised in the challenge header.
The default values is "" which omits the algorithm field in the challenge header and uses MD5 for digest authentication.