auth: add support for Authentication-Info header
Add an option to include an Authentication-Info header in the response to a successful authentication.
The header can be enabled with the parameter 'add_authinfo_hdr' module parameter.
The implementation follows RFC2617.
The Authentication-Info header is added inside the function post_auth of the auth api.
The parameter HA1 had to be added to the post_auth api function which requires an
update in all auth_* modules that use the module auth api.