registrar: Check max URI size of contact also for first contact header
The URI size of contacts is checked against a max size to ensure that it fits into the database column.
This check was already present before but was done only for the contact header fields other than the first header field
which made that REGISTER with a single contact were not checked for max size and were truncated later when stored to the database.
With this fix all contact URIs of all contact header fields are checked against the max size.

(cherry picked from commit c0697d76fc2ae8ea9676aa9a9b57db442b9c678f)