registrar: Add module parameter "contact_max_size" to make max contact size configurable
The new module parameter "contact_max_size" allows changing the max size of contact URIs
that are accepted in REGISTER requests.
The default value of this parameter is 255 which was the value of the compile time constant
that is replaced by this module parameter.
If configured one must make sure that the DB actually supports the configured size in the
column "contact".