ims_usrloc_scscf: Double check contact refs on DB
If the contact is on state CONTACT_DELAYED_DELETE
and still has refs > 0 in memory, it could be
that it failed to acquire a lock to de-ref the
contact. This brings to a leak on the contact DB
because the contact will always have refs > 0 and
no actual refs on DB. Now, as a fallback, we
actually checks on DB the real number of refs of
the contact. If this refs is 0, force the
number  of refs in the record to 0 and mark the
contact for deletion.