Added some additional debug statements to the code
I needed to add some additional debug statements to the code in certain
areas due to the issue with the seize line on calls. Even though it does
not seem to be giving me all the debug statements in the log file, I do
know that the code is executing. Also, it maybe because some of the
LM_-based statements did not end with a line terminator.