Caller ID follows the seize after being held.
It states in the code of sca_call_info_is_line_seize_reinvite that if
the from and to URI's are the same, then a reINVITE is occuring. Well
this was taking any 200 reply and changing the owner and appearance uri
to the wrong things. This also cleared up outbound calls to be placed
and hold and then seized by another phone without multiple appearances
showing up.