sca: Fix the CID on seizing a held call from a different SCA AOR UA.
- Part of the issue reported by Kevin Scott Adams, GH #341. On seizing a
held call from another SCA subscriber UA, the Caller ID becomes the
actual AOR of the SCA. In reviewing the SIP traces the INVITE and REPLY
of seizing a held call as the 'from' and 'to' URI's set to the AOR. In
the sca_call_info_is_line_seize_reinvite() function this is handled but
was not handled when a 200 reply was issued. Uploaded, compiled and
tested.