rr: use default port for protocol if it is omitted in route header
When using double route headers the selection of the outgoing
socket is only done on IP address if the port is omitted in that
route header. This fix defaults the port to 5060/5061 depending on
protocol so the correct listen socket is chosen. When no socket is
found it will fallback to the current behaviour keeping backwards
interoperability.

(cherry picked from commit aaec127fbf1f3be41e49601330d7eafe7353c53a)