sca: sca_call_info.c: Remove redundant code
- In reviewing the code block that I removed it is actually duplicated
in the sca_create_cononical_aor_for_ua() function which is called from
sca_create_canonical_aor() which is called inside this 'if' block, so
why do it twice. The function sca_get_msg_contact_uri() is also called
in sca_create_cononical_aor_for_ua() but it does not set or pass a
'contact' parameter so we still need it in sca_call_info_update() just
don't need to check for errors because it will error out in
sca_create_cononical_aor_for_ua() and do the same thing.
- Maybe in the long run for the 'contact_uri' parameter, create a second
function or add it as a setable paramter to the
sca_create_cononical_aor_for_ua().
- Code change has been tested and running for about 2 months without
issues.