Add comments to db schema and default timestamp
Added comments explaining table functionality to MySQL schema. These comments can be viewed in MySQL when browsing the schema which helps admins understand what each table is used for.

Changed default value for "last_modified" in grp table from "1970-01-01 00:00:00" to CURRENT_TIMESTAMP and set to automatically update "last_modified" if the record is ever updated. This means that if a record is inserted without a "last_modified", MySQL will insert an accurate time-stamp instead of 0000-00-00 00:00:00 which breaks the SQLOPS module.