app_python3: implement script reloading
Copy the method used in app_lua: the version of the script is incremented by RPC.
In python_exec.c:apy_exec() check for a newer version and reload the script.

Set a lock so the script reload only occurs at depth 0 (in the unlikely case
that apy_exec() is called recursively).

This is not thread-safe as we are using a process-wide lock: don't call back
into apy_exec() from a Python extension that uses threads.