packaging: Fix typo in kamailio.spec
There is a number of occurrences of `%bcond_with|without rebbitmq`. Since the condition to include the `rabbitmq` package (line 633) is `%if %{with rabbitmq}`, this conditional statement would not have the intended effect. If this was not a deliberate typo for some reason, this commit fixes the issue by changing all `%bcond_with|without` occurrences of `rebbitmq` to `rabbitmq`.