pkg/kamailio/deb: also restart service if unclean exit code
New:
If set to on-failure, the service will be restarted when the process exits with a non-zero exit code, is terminated by a signal (including on core dump, but excluding the aforementioned four signals), when an operation (such as service reload) times out, and when the configured watchdog timeout is triggered.

Old:
If set to on-abort, the service will be restarted only if the service process exits due to an uncaught signal not specified as a clean exit status.

https://www.freedesktop.org/software/systemd/man/systemd.service.html

Signed-off-by: Kevin Olbrich <ko@sv01.de>