pkg/kamailio/oracle: wait for network to be up
Units that strictly require a configured network connection should pull in network-online.target (via a Wants= type dependency) and order themselves after it. This target unit is intended to pull in a service that delays further execution until the network is sufficiently set up. What precisely this requires is left to the implementation of the network managing service.

https://www.freedesktop.org/software/systemd/man/systemd.special.html
https://lists.kamailio.org/pipermail/sr-users/2018-November/103668.html

Signed-off-by: Kevin Olbrich <ko@sv01.de>