kamctl: keep it POSIX
Use portable "#!/bin/sh" shebang. This may still be a symlink to bash
(or some other shell).

Also replace the redirection to "/dev/stderr" with a POSIX redirection
to stderr. "/dev/stderr" is not available on all systems and is also not
recognized by all shells (example: ash). On systems without a working
"/dev/stderr" the user will not see the prompts etc.

Signed-off-by: Sebastian Kemper <sebastian_ml@gmx.net>