core/dset: Create Contacts header with path vector as Route header
When Kamailio is functioning as a registrar sends a 302 redirect, the
registered contacts are sent inside the Contact header. If a contact
as a path vector, that path should become a Route header in the
INVITE (rfc 3327). This patch adds the path vector as a Route header
parameter to the Contacts in the redirect so the receiving entity of
the 302 can add the proper Route headers.

This is a standard way to specify headers to be added to a request,
see rfc3261 section 19.1.1.