ims_ipsec_pcscf: IPv6 support
The parameter ipsec_listen_addr of the module can be set to either IPv4
or IPv6 address. All xfrm related code is reworked to handle both type
of addresses.

Note: At the moment it is not possible to use both IPv4 and IPv6 at the
same time for IPSec. The implementation allows it, but additional config
parameters should be added in order to make this work.