imc: make behavior of "#create" configurable when room already exists
The default behavior of the create command is to silently add the user
to a group of one already exists. This can be confusing in some
scenarios, e.g., when the user miss-types the room number. This patch
adds a module parameter called "check_on_create" which makes it
possible to customize this behavior. If set and the user attempts to
create a room that already exists, the command will report an error.