imc: make #create command arguments optional
With no arguments, the room name is taken from To header and the room
will be made public.

With one argument, if the value is 'private', a private room with name
from To is created. If the value of the single argument is anything
else, it is taken as room name and the room will be made public.

With two argument, the first argument is room name and the second
argument must be string 'private'.

This change makes the #create command consistent with all other
commands where the room/user name is also optional.