lcr: enhance the lcr.dump_rules with filtering params:  - lcr_id [optional]  - prefix [optional]
If lcr_id param is provided, then only the rules belonging
to the given lcr_id will be dumped.
If the lcr_id and prefix param is provided, then only the rules
belonging to the given lcr_id and prefixes that equal or longer
to the given prefix will be dumped.