More string buffer handling improvements in imc module
Let the compiler do the work of figuring out the size of the string
buffer being written to using the sizeof operator. Hopefully, this
will make the source code less error-prone and more Coverity friendly.

Properly handle all return values of snprintf. In particular, do not
fail silently if the function returns -1.