core: switch to not use SIG_DEBUG by default
The code enabled by SIG_DEBUG inside signal handlers is
async-signal-unsafe (see the signal-safetey(7) man page), as it calls
at least syslog(3) and stdio functions.

This can crash the daemon, and while this is certainly not easy to trigger,
it's still not correct to leave enabled in production builds. It might be
an acceptable tradeoff as a debugging aid, though.

So we switch the code to off by default, and swap the meaning of the
macro to enable the unsafe debugging, to match.