tm: new weight-based call-termination distribution
Extended t_load_contacts function with the parameter 'mode'.
It can be used to define the algorithm to use for ordering
the contacts.
If 0, or missing, the standard algorithm with be used.
If 1, the new weight-based algorithm will be used.
The new algorithm uses the q value as the weight of the branch.
Higher is the q value, higher is the probability to be tried
as first. Equals q values means equals probability to be tried.
Branches with q values equals to 0 or lower are not considered by
the ordering algorithm, but just added at the end of the list
as backup option if all other branches fail.