dispatcher : better distribution when using hash and destination is not available.
When a destination is selected using one of the hash algorithms (over Call-ID, From, etc.)
and destination is not available, then the next is selected. This causes the following problem :
All messages to the unavailable destination will be sent to the same unique destination.
This will cause that one destination will receive double load and the rest will continue with the same.

Solution : To better distribute messages, while also guaranteeing the same distribution of the hash, a re-hash of the hash is made.

As the rehash could select the same or other failed destination, it can rehash several times. The number of rehashes made is controlled by :

modparam("dispatcher", "ds_rehash_max", -1)

Which can take the following values :

-1 : The total count if destinations is used 
0 : Disabled. Reverts to original implementation.
Positive number. Maximum number of rehashes.