core: add sip_parser_log core variable
* when handling non sip messages, parser errors are logged at core level
which defaults to ERR. since this may induce in error sysops or log
scanners, sometimes the core level is set to debug just to omit these
messages from log, which is not good since other messages are omitted
too. this commit adds a new core variable to set the log level for
message parsing errors. this way, we can explicitly set the log level
for message parsing while leaving the core level at ERR.